+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Compatibility
 * `Hash#merge` now preserves the key order from the original hash for merged values (#1650).
 * Coerce values given to `FFI::Pointer` methods.
 * `FrozenError` is now defined and is used for `can't modify frozen` object exceptions.
+* `StringIO` is now available by default like in MRI, because it is required by RubyGems.
 
 Changes: